Added the input coordination to the logic
