+++ src/inputLayer/MovementOptions.java
@@ -1,3 +1,4 @@
+package inputLayer;
 import java.util.Scanner;
 
 public class MovementOptions {
